Joysticks support improvement, including wiimote controller
the wiimote controller sends hat1, hat2, hat3 events.
No hat0 events.
SDL2 assumes that if there are 3 hats, it must be from 0 to 2.
I inform the wii driver coder of this issue.
I modified SDL2 to not assume this point while some other joysticks
could have the same issue.

For the moment, it's wiimote or wiimote controller, not both at the same time,
because these have the same sdl guid.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis-lamarre@gmail.com>