odroid xu4 : sdl2
upgrade sdl2 from 2.0.2 to 2.0.3 on the odroid
I created the patch by geting the delta between the sdl2 odroid branch and the official sdl2 and cleaning it.
and i applied this patch over 2.0.3 while it's not intrusive.
I mainly add a video driver and a switch in the configure script.
Note that this patch is applied for the rpi too, but the switch is not applied because the check fails.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis.lamarre@gmail.com>