joystick: fix the wiimote controller extension with sdl 2.0.4
in sdl 2.0.4, if ID_INPUT_KEY and ID_INPUT_JOYSTICK are defined, a linux event (/dev/input/*) is both a joystick and a keyboard.
so, for 1 linux event, sdl produces 2 events.
At the end, es sees 2 events.
In my case, the wiimote button value is the arrow bottom keys.
At the end, es go down 2 times in the lists for 1 button on the down button.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis.lamarre@gmail.com>