reicast: fix reicast
note that reicast requires OpenGL 3.1.
My computers have only 3.0, so i get an error : GLXBadFBConfig and can't test.

To get your version :
glxinfo | grep "OpenGL version string"

Note that it would be nice to be able to display a correct error message to the user instead
of just going back to es.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis.lamarre@gmail.com>