Fixup rockpro wifi and bluetooth
Please note that in my own builds I have a very custom kernel with a very different device tree and a huge amount of commits back-ported to 4.4 from mainline. Many of those are changes to the wifi, sdio, bluetooth, and uart drivers. So I have specifically written this patch based off the kernel commit you're linked to and I believe it should be all that's required to enable these features. You'll want to test their functionality and stability without all of my other customizations.