Avoid duplicate resolution lines when using optimus
Nvidia Optimus causes the OS to see two screens - one on the Intel adapter which is used for display output and one on the Nvidia adapter which is used for rendering the frames. So we have two lines returned from xrandr --currentResolution and one must be discarded. ES will not launch otherwise.