fix pcsx2 memcards
pcsx2 saves column information in the .ini file with the
translated names, not the original names.
This add latin strings to the .ini causing troubles to
configgen that consider the file as a utf-8 one.
pcsx should not used translated strings for technical
variable imho. so, i just remove them. There are just
the columns width in the gui.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis.lamarre@gmail.com>