change RPCS3 SPU decoder to Recompiler (LLVM)
This seems to function perfectly fine and is the default of RPCS3 itself, there's only a handful of games that even need this to be different.
Considering that this dramatically improves performance on all hardware it should be the default. I will investigate implementing this as an option into ES itself in the meantime.