add 'resize_tattoo' option to tattoo 2
Second iteration of https://github.com/batocera-linux/batocera.linux/pull/5161

"original" = original size of the image (no resize)
"240/1920" = resize proportionally to 240th of a 1920 px screen
"225/1920" or "auto" = resize proportionally to 225th of a 1920 px screen
The reason it's "#/1920" is because it's proportional to the bezel, not the screen itself, and our included bezels are 1920. This also includes if the user chooses to stretch the bezel (hopefully) or if another resolution is being used altogether.

Complements https://github.com/batocera-linux/batocera-emulationstation/pull/1063