use core provided as default aspect ratio for libretro
By default, RetroArch will use a 4:3 aspect ratio when no aspect ratio index is provided. This is what happens in Batocera when the user leaves the "GAME ASPECT RATIO" setting on "AUTO". This is suitable for most systems.

However, systems such as the Game Boy and various vertical Arcade games that use different aspect ratios outside of 4:3, and look weird and distorted when forced to 4:3.

Note that this change will not affect anything when decorations are turned on, as they use their own viewport settings. This is only for when decorations are set to "NONE".