lr-mame Fix
lr-mame will read lightgun buttons as joystick buttons, causing duplicate button mapping. As this seems to be the only core to do this, I added a "map lightgun to controller" option. If left on Auto, it will behave as normal, except if the mame core is used. The mame core will skip mapping the lightgun buttons. It can be manually set to on and off per system or game if needed.