Add a bunch of RPCS3 options to ES
Unfortunately the 4:3 ratio in RPCS3 seems to be glitched: instead of showing the 4:3 aspect ratio at its intended resolution, it will instead stretch out the display to whatever your display actually is, irrelevant of whether "stretch to display area" is enabled or not. Still, implementing this is better than not implementing it, as there are some users on genuine 4:3 screens who would appreciate this option.