increase audio latency by 32ms
Avoids audio crackling in most systems that is present (though not completely obvious unless you're specifically listening for it).
The raspberry pi 3 is not as powerful as the other SBCs and should probably have a slightly higher buffer anyway. 32ms increased delay is nearly unnoticeable and can always be turned back down by the user.