bugfixes + feature implementation
Although it's possible to get the config reader to properly put in the codes on the first launch, any subsequent reads cause the cheat files to become corrupted. Therefore, I've resorted to using a more robust approach: just copy the data in if it detects the files as missing. This is the easiest to understand while still allowing users to implement custom codes if they want to. If they rekt the cheat files, then simply deleting them will restore everything.