allow shift+f1 to open RA quick menu on keyboard-only
Fixes the chicken-and-egg issue of opening up the RetroArch menu to allow the user to rebind their keyboard controls when only a keyboard is available.

Existing users will not have their config changed, they will still have Esc be their "immediately close RetroArch" button. But, new users or users who erase their RetroArch custom config will have to press Shift+Esc to exit. Considering the keyboard is not mapped consistently between emulators anyway, I don't see this as an issue. If it is, though, we could easily add in an evmapy keys profile to re-enable the exiting immediately upon pressing Esc functionality (in my opinion, having a single button exit the emulator is not a good practice in the first place, especially when saves are at stake).