[v35] Update some .format() strings to f-strings part 1
This results in an improvement to the launching times of games. This is a new feature to Python 3 which improves the legibility of the code and increases its efficiency in execution. Since we made the jump to Python 3, we may as well take advantage of it.

I've run only the configgen folder this through `pyupgrade` with its minimal settings in order to reduce the chances of this procedure dramatically breaking something. I then manually went through every file to ensure that they all were still okay.

I've tested most systems that this affects and they all seem to be functioning fine, however this is something that cannot be thoroughly tested without sending it out to the masses. So please wait for v35's beta to merge this.