Let kernel decide the default cpu governor policy per core
Prevent S18governor from setting the governor on all profiles by default. This seems to be unintended behaviour.

From what I gather from the intention of https://github.com/batocera-linux/batocera.linux/pull/3969 , this was originally meant to just set the governor on all cores if the user has set a particular profile. If there is no setting, this script will grab the currently set one (the one requested by the kernel, which should be the "default" one) and then save that value to `batocera.conf`'s `system.cpu.governor` key.

The problem comes from the fact that on next boot, this script will check for that setting, see that it is set to whatever value it is that it grabbed from the last boot, *and then apply it to all the cores*. This will override whatever kernel requested default there may have been for each individual core (many kernels opt to make one CPU the "main" one running at full power, and leave the rest on demand if they are needed) resulting in unintended default behaviour.

By default a script like this should not be setting kernel settings without the user's consent, it should leave it as it is and only act if the user has explicitly set it. This PR removes the behaviour it has of adding the key into `batocera.conf` if it is not found, this solves the behaviour.

This change may explain why people have been complaining about Batocera overheating their SBCs, and why handheld users are experiencing much lower battery life on Batocera compared to other OSes. I've asked a few users to test different settings, as well as an absence of it by commenting out the code. Will leave as draft until there is more concrete evidence.