increase mouse.patch default deadzone to 0.2
This should decrease the amount of people complaining about the on-screen cursor moving by itself when using pad2key's emulate mouse-cursor feature.

I will look into making this a configurable variable in the future. But that would involve a lot of work in regards to standardising deadzones, so it may not be for a long time.