fix win600 splash + add "batocera-resolution getDisplayMode"
Adds "getDisplayMode" to `batocera-resolution` such that it can be accessed this early in the boot stage. May also help with rotational things in the future.

Using this, the splash screen now also respects the rotation setting set by the user. Fixes the bugged rotation of the splash screen on win600.