bump xemu to 0.7.67
Includes updates to the configgen and es_features to accomodate for changes within the emulator. Notably, Xemu will now reject the config file if there are any options with empty values in them, something the old configgen for this system relied on. All of those instances had to be removed. Tested and working fine.

Also took liberty to add "Enable DSP audio" as a user editable option, instead of just hard-coding it to be off.

Networking should work out of box, but I won't be able to test for that for a while.