DS Mouse Fix, Part 1
This enables an option for mouse (or touchscreen) input for the touch screen in MelonDS. Previously we were forcing joystick, MelonDS' touchscreen controls are mutually exclusive.

Part 2 will be if I can fix Retroarch mouse assignments - currently if you have a mouse and a Sony controller connected, the mouse will be assigned to port 1 and the touchpad to port 2, so you can't use the touchpad in Retroarch. I'll be looking to see if there's a workaround for that.