Tweaking Sinden Lightgun settings to recommended
Default should be Exposure -7, Contrast 100, Brightness 50. ExposureAuto is very bad when light in gun games fluctuates. Sinden Lightgun community and testers recommend Exposure set to -7, Contrast to 120 and Brightness to 60. I've tested countless and it works best for thinner borders with high light. Camera is losing track in current values.