Set SDL_GAMECONTROLLERCONFIG when launching ports
Most Linux ports are written in SDL, therefore it is extremely useful to
export the (known working) controller mappings as it is done for other systems.
This is invaluable for older binaries with their own copy of SDL and
controller mappings, as they won't work out-of-the-box with new
controllers and/or kernel drivers.

This does not have any impact on non-SDL ports.