Fix swtrilgy supermodel light gun
That value was introduced  in v37 beta which breaks the light gun support for that specific game. Without that added value, the light gun works again. This value is not needed and break the light gun patch for the game (it is an analog_joystick instead of analog_gun game).