Update vpinball.keys
The right and left magnasaves are not currently mapped on the controller. Haunted House is unplayable without that.
It makes sense to map them to the triggers (as the flippers are already mapped to the buttons).
This way, tables needing magnasaves will be playable right out of the box with a x360 controller (for example).