Compress lr-mame xml DATs with gzip, fixes #8556
Disk space in datainit is at a premium, as the fallback ramdisk has a
total of 256M available.

Around 20% of this space is consumed by lr-mame 2003 and 2010 XML DAT
files.  While these files are handy to have in the distribution, they
are not directly used by the installed system.  A significant chunk of
space can be saved by compressing them with gzip at the highest
compression ratio.  Users can easily decompress them if needed for
reference or to validate a romset.