add game id for lindbergh
Some games are missing from the MAME database, but other do exist. Since many games rely on our gamesdb.xml to be detected as light gun or wheel games, I wanted to be futureproof by adding more information in the document.

I've added the missing games in gamesdb.xml here: https://github.com/batocera-linux/batocera-emulationstation/pull/1851

With this, the working ROMset will properly be detected in ES, in gamesdb.xml and in scrapers.