liretro: fix modern arcade stick alt layout
The current implementation in lr-mame with mapping through .cfg
is not (or not work anymore) with trigger axis for L2 R2.

That break modern arcade stick alternative layout.

This patch is fixing, doing the switch of L1 R1 -> R1 R2 in the retroarch
remap directly. (and the classic layout is still done in lr-mame cfg)

Also add support for fbneo && flycast core