always write controller mouse index
This comment about lightgun plugged is wrong.

the "lightgun" param, is about lightgun_mapping option,
to map lightgun games to controllers.

The true lightgun is handled later, and overwrite the mouseindex again if present,
so no conflicy is possible.

currently, lightgun_mapping is True by default, so mouseindex is never written