batocera-wine: Errorcontrol and restored stopFunction
we have a better control over the instances of wineservers now

The waitWineServer function need two arguments now, time and the PID of the process to watch for.

So `waitWineServer 0 $(pgrep -f "${WINESERVER}")` is the standard calling function now.

The problem is that there always 2 instances of wineserver at least. With the introduction of the trap wineStopServer it isn't clear anymore which instance is the current one. So I passed the `${WINESERVER} -k` background PID to the waitWineServer and here it reports if the PID is valid or not.

I think that will also help to hunt down future changes