scummvmGenerator: get game ID from file content
In favour of https://github.com/batocera-linux/batocera.linux/pull/13922

Tested and method works with libretro and standalone scummvm. 

Just a small addition... we can add files to scummvm and call them `whatever.scummvm` and just set the game ID inside the file.

I added a handler that files less then 3 bytes are getting ignored and handled like usually (see ... a "empty" windows file with with CRLF is 2 bytes long) ... afaik the shortest game ID is `11h` and this is 3 bytes ;)  So we should have a perfect match here.

An example....
Create a file `Goblins III [DOS 1.02 CD French].scummvm` inside the Goblins3 games folder. Edit it and add `gob3-cd-fr` as ID inside.... isn't that nice?

The great **plus** is that we can achive compatibility with Windows ScummVM for example ... I really struggled about the save-game names -- they are built from the ID-name!