scummvmGenerator: Both cores work the same way
I've tested:

**SCUMMVM-file:**
1. `Whatever you named this.scummvm` -> empty
- standalone uses autodetection
- libretro uses autodetection

2. `Whatever you named this.scummvm` -> contains valid gameID `monkey2-cd-fr` or engineID `monkey2`
- standalone uses the content inside file
- libretro uses the content inside file

3. `engineID.scummvm` -> empty
- standalone uses filename
- libretro uses filename

4. `engineID.scummvm` -> contains valid gameID `monkey2-cd-fr`
- standalone uses the content inside file
- libretro uses the content inside file

5. `gameID.scummvm`-> empty
- standalone uses the filename and can set valid language automatically
- libretro will fallback to autodetection

**SQUASHFS-file:**
behaviour is exactly like SCUMMVM-file except:
- There is no scummvm-file available inside the sqashfs-file
  - standalone uses autodetection
  - libretro uses autodetection

In general:
If the user fills a file with content, this has top priority and is accepted without being checked. This is the natural behaviour of libretro-core and would also affect standalone-core.
This PR-commit adds the automatic detection for games also for the standalone-core. This makes a simple setup for the user possible. Especially for squashfs files, which no longer need to contain a scummvm file.

I would like to see this featurea in version 42.

PS: This wasn't done with help of AI just HI/NI