add cpu slider for hotd4 (lindbergh)
At default/auto, it keeps the formula for minimum frequency. If the user changes the slider in ES, then it takes the value chosen between 0.5 to 4.0 (moving 0.1 each). Idea is to select base frequency. I have 3600MHz processor, I need to go with 3.6 for perfect speed accuracy on this game.