Update batocera-xtract: open archives, extraction of single files, creation of archives
1. We can open archives (7z, rar, tar and zip) and we can extract files out if them with pcmanfm (you have to write command line `file-roller --open` and save the application with a usage name. It will open a yad dialog, from there you can extract files.
2. Extraction of _zip 7z rar tar gz xz tar.gz tar.xz_ with pcmanfm **context menu**
2. We can create compressed zip files from context menu now. Only zip is supported

----

If you want to help then please add some more archivers. I tried to add more file-info (size, file/dir, time and date) but this ended in an endless battle of timing. I needed 25s to open a zip with 300 files. Now it just takes 1 or 2 seconds.

-----

To add the open function... go to command line and enter `file-roller --open` for each type of archive. Currently only **zip, 7z, tar and rar** can be browsed by filemanager. But it can be extended if you have some coding skills. I avoided to have fancy useless things..... You have to tick the default option and you have to name the application.

This data is stored to `/userdata/system/.config/mimeapps.list` so it is the defalut action for the setted MIME type

-----

I also added the compress option. Currently only zip is supported, so select the files and click **Compress on the context menu**, the file is stored into the parent directory of your selected file(s).... also directories are supported.

----

Greetings fly out to the whole Batocera-dev-Team for this nice piece of software. It's pleasure to see things working!