Providing a workaround to specify autoexec section in dosbox-x (#14491)
While auto-mounting the game directory as C: in dosbox-x games works for the majority of users, some expert users may find it restrictive. Methods like adding commands to a dosbox.bat script may fail if, e.g,  you want to unmount C: to mount an image instead.

I propose an expert method that can be used if the current standard is not useful for the user in a given situation: to use a dosbox.aut file created in the game directory (which must include the [autoexec] label) that will be read and appended at the end of the dosbox-custom.conf file. No other commands will be run instead of the ones in this file for this use case.